# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM Pay API",
        "description": "API pour l'application OM Pay - Gestion des paiements et transferts",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        },
        {
            "url": "https://om-pay-mk4h.onrender.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/{numeroCompte}/paiement/scanner-qr": {
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Scanner un QR code",
                "description": "Scanne et décode un QR code de paiement marchand",
                "operationId": "4e9b0413d6ee91b924428e8d65890a86",
                "parameters": [
                    {
                        "name": "numeroCompte",
                        "in": "path",
                        "description": "Numéro de compte de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "7735434534"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "donneesQR"
                                ],
                                "properties": {
                                    "donneesQR": {
                                        "description": "Données encodées dans le QR code",
                                        "type": "string",
                                        "example": "QR_CODE_DATA_HERE"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "QR code scanné avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "marchand": {
                                                    "type": "object"
                                                },
                                                "montant": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "reference": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "QR code invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "QR code invalide"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/{numeroCompte}/paiement/saisir-code": {
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Saisir un code de paiement",
                "description": "Saisit manuellement un code de paiement marchand",
                "operationId": "b03402b3f9fcf0bceecf7221319f98bd",
                "parameters": [
                    {
                        "name": "numeroCompte",
                        "in": "path",
                        "description": "Numéro de compte de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "7735434534"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "montant"
                                ],
                                "properties": {
                                    "code": {
                                        "description": "Code de paiement fourni par le marchand",
                                        "type": "string",
                                        "example": "PAY123456"
                                    },
                                    "montant": {
                                        "description": "Montant du paiement en XOF",
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Code de paiement validé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "idPaiement": {
                                                    "type": "string",
                                                    "example": "OM20251111131953ABC123"
                                                },
                                                "marchand": {
                                                    "properties": {
                                                        "idMarchand": {
                                                            "type": "string"
                                                        },
                                                        "nom": {
                                                            "type": "string"
                                                        },
                                                        "logo": {
                                                            "type": "string",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "montant": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 5000
                                                },
                                                "devise": {
                                                    "type": "string",
                                                    "example": "XOF"
                                                },
                                                "dateExpiration": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "valide": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Code de paiement invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Code de paiement invalide"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/{numeroCompte}/paiement/{idPaiement}/confirmer": {
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Confirmer un paiement",
                "description": "Confirme et exécute un paiement en attente",
                "operationId": "2693b3eba5adcd24e7d5b1f3b0738d57",
                "parameters": [
                    {
                        "name": "numeroCompte",
                        "in": "path",
                        "description": "Numéro de compte de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "7735434534"
                        }
                    },
                    {
                        "name": "idPaiement",
                        "in": "path",
                        "description": "ID du paiement à confirmer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "codePin"
                                ],
                                "properties": {
                                    "montant": {
                                        "description": "Montant du paiement (optionnel pour vérification)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 2000
                                    },
                                    "codePin": {
                                        "description": "Code PIN de l'utilisateur",
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paiement confirmé et exécuté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "idPaiement": {
                                                    "type": "string"
                                                },
                                                "statut": {
                                                    "type": "string",
                                                    "example": "REUSSI"
                                                },
                                                "dateExecution": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "PIN incorrect ou paiement déjà traité",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}